Remove QR code type check since it was rejecting valid streaming QR codes.
ResultParser.parseResult will try to parse the decoded QR result into URI, SMS or other type.
It's default when the parsers fail, is to fall back to TEXT type.  Since we are encoding byte
data in the QR code and not characters, sometimes the parser will interpret the data encoded
in the result as a URI and then fail to identify it as part of a streaming QR.  Remove this
type check for now but try handling this later whenever the algorithm identifies and improperly
formatted QR code and then re-interpret the result as if it were a single standard QR code.